Stop clamping stack merges, and pay out whole
Dropping 50,000 gold onto another 50,000 left one pile of 65,535 and destroyed
the rest: both merge paths folded the amounts with a `saturating_add` on the
`u16` an `Amount` is stored in, then despawned the source. The engine's first
item-loss bug.

The cap is explicit now — `items::MAX_STACK`, 60,000, ServUO's `Item.WillStack`
number, kept clear of the `u16` edge — and the overflow goes back to the player
rather than nowhere. The references disagree on how: ServUO refuses a merge whose
sum would pass the cap, so a full pile simply will not take a drop, while Sphere's
`CItem::Stack` fills the destination to its maximum and leaves the remainder on
the source. Sphere's is the kinder of the two and loses nothing either way, so
`merge_amounts` fills what it can and reports what is left, and `merge_onto`
bounces a remainder home instead of despawning it.

Where the world hands goods over rather than a player, `items::give` now spreads
what it is given across every matching pile with room and starts as many new ones
as the rest needs — a container ends up holding two gold piles, as in UO. It takes
a `u32` for the same reason `take_from_container` does: a large sale earns more
than one pile holds, and the vendor's `earned.min(u16::MAX)` was paying 65,535 for
a 100,000 sale and saying nothing about the difference.

+++ crates/world/src/tick.rs
@@ -984,7 +984,7 @@ impl World {
             return;
         };
         if stackable {
-            items::give(&mut self.state, backpack, graphic, hue, amount);
+            items::give(&mut self.state, backpack, graphic, hue, u32::from(amount));
         } else {
             items::place_one(&mut self.state, backpack, graphic, hue, amount);
         }
